+++ panda/src/display/touchInfo.cxx
@@ -15,40 +15,40 @@
 #include "touchInfo.h"
 
 TouchInfo::TouchInfo(){
-	_x = 0;
-	_y = 0;
-	_id = 0;
-	_flags = 0;
+  _x = 0;
+  _y = 0;
+  _id = 0;
+  _flags = 0;
 }
 
-LONG TouchInfo::get_x(){
-	return _x;
+int TouchInfo::get_x(){
+  return _x;
 }
 
-LONG TouchInfo::get_y(){
-	return _y;
+int TouchInfo::get_y(){
+  return _y;
 }
 
-DWORD TouchInfo::get_id(){
-	return _id;
+int TouchInfo::get_id(){
+  return _id;
 }
 
-DWORD TouchInfo::get_flags(){
-	return _flags;
+int TouchInfo::get_flags(){
+  return _flags;
 }
 
-void TouchInfo::set_x(LONG x){
-	_x = x;
+void TouchInfo::set_x(int x){
+  _x = x;
 }
 
-void TouchInfo::set_y(LONG y){
-	_y = y;
+void TouchInfo::set_y(int y){
+  _y = y;
 }
 
-void TouchInfo::set_id(DWORD id){
-	_id = id;
+void TouchInfo::set_id(int id){
+  _id = id;
 }
 
-void TouchInfo::set_flags(DWORD flags){
-	_flags = flags;
+void TouchInfo::set_flags(int flags){
+  _flags = flags;
 }
